[warning-fix][implicit-conversion-from-wider-type]
Fixed various cast warnings

+++ src/Draw/CommandWindow.cxx
@@ -93,7 +93,7 @@ LONG APIENTRY CommandProc(HWND hWnd, UINT wMsg, WPARAM wParam, LONG lParam )
     			hWndEdit = (HWND)GetWindowLong(hWnd, CLIENTWND);          
     			MoveWindow(hWndEdit, 0, 0, LOWORD(lParam), HIWORD(lParam), TRUE);
             // Place le curseur a la fin du buffer
-          index =  SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
+          index = (int) SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
           SendMessage(hWnd, EM_SETSEL, index, index); 
     			break;
 
@@ -103,7 +103,7 @@ LONG APIENTRY CommandProc(HWND hWnd, UINT wMsg, WPARAM wParam, LONG lParam )
           break;
 
     default :
-					return(DefWindowProc(hWnd, wMsg, wParam, lParam));
+					return LONG(DefWindowProc(hWnd, wMsg, wParam, lParam));
   }
   return(0l);
 }
@@ -164,7 +164,7 @@ int GetCommand(HWND hWnd, char* buffer)
   int again = 1;
   char temp[COMMANDSIZE]="";
 
-  int nbLine = SendMessage(hWnd, EM_GETLINECOUNT, 0l, 0l);
+  int nbLine = (int)SendMessage(hWnd, EM_GETLINECOUNT, 0l, 0l);
   
   int nbChar = 0;
   buffer[0]='\0';
@@ -175,7 +175,7 @@ int GetCommand(HWND hWnd, char* buffer)
       WORD* nbMaxChar = (WORD*)temp;
       *nbMaxChar = COMMANDSIZE-1;
       
-      int nbCharRead = SendMessage(hWnd, EM_GETLINE, nbLine-1, (LPARAM)temp);
+      int nbCharRead = (int)SendMessage(hWnd, EM_GETLINE, nbLine-1, (LPARAM)temp);
       nbChar += nbCharRead ;
       int cmp = strncmp(temp, PROMPT, 10);
       temp[nbCharRead]='\0';
@@ -226,16 +226,16 @@ LONG APIENTRY EditProc(HWND hWnd, UINT wMsg, WPARAM wParam, LONG lParam )
 									strcpy(console_command, buffer+strlen(PROMPT));
 									console_semaphore = HAS_CONSOLE_COMMAND;
 									  // Purge du buffer
-                  nbline = SendMessage(hWnd, EM_GETLINECOUNT, 0l, 0l);
+                  nbline = (int)SendMessage(hWnd, EM_GETLINECOUNT, 0l, 0l);
 									if(nbline > 200)
 									{
                       nbline = 0;
 											GetCommand(hWnd, buffer);
-                      index = SendMessage(hWnd, EM_LINEINDEX, 100, 0);
+                      index = (int)SendMessage(hWnd, EM_LINEINDEX, 100, 0);
 											SendMessage(hWnd, EM_SETSEL, 0, index);			
 											SendMessage(hWnd, WM_CUT, 0, 0);
                         // Place le curseur en fin de text
-                      index =  SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
+                      index =  (int)SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
                       SendMessage(hWnd, EM_SETSEL, index, index);                      
 									}
     				      return(0l);
@@ -244,7 +244,7 @@ LONG APIENTRY EditProc(HWND hWnd, UINT wMsg, WPARAM wParam, LONG lParam )
                   if (IsAlphanumeric((Standard_Character)LOWORD(wParam)))
                   {
                       // Place le curseur en fin de texte avant affichage
-                    index =  SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
+                    index =  (int)SendMessage(hWnd, WM_GETTEXTLENGTH, 0l, 0l);
                     SendMessage(hWnd, EM_SETSEL, index, index);
                     CallWindowProc(OldEditProc, hWnd, wMsg, wParam, lParam);                    
                     return 0l;
@@ -256,7 +256,7 @@ LONG APIENTRY EditProc(HWND hWnd, UINT wMsg, WPARAM wParam, LONG lParam )
     if (console_semaphore != WAIT_CONSOLE_COMMAND) 
       return 0l;									
   }
-  return CallWindowProc(OldEditProc, hWnd, wMsg, wParam, lParam);
+  return (LONG)CallWindowProc(OldEditProc, hWnd, wMsg, wParam, lParam);
 }
 #endif
